fix: resolve all Sonar issues across new modules
Addresses all Sonar findings:

Blocker bugs:
  - dir.c: out-of-bounds access on empty path (tmp+1 when len==0)

Critical complexity reductions:
  - csv.c: extracted grow_buf, buf_append, read_quoted, read_unquoted,
    grow_fields helpers to reduce cognitive complexity 76 -> under 25
  - glob.c: extracted match_char_class helper to reduce complexity 56 -> under 25
  - semver.c: cognitive complexity 27 -> under 25

Code smells:
  - logger.c: &default_sink, memcpy instead of strcpy, const tm_info
  - io.c: removed dead store to eof (2 occurrences)
  - semver.c: removed redundant (int) cast in parse_num
  - uuid.c: removed redundant (int) casts in hex_val
  - thread_pool.c: removed redundant (size_t) cast
  - spsc.c, test_thread_pool.c: <Windows.h> -> <windows.h> (lowercase)
  - benchmark_csv/semver/uuid: pointer-to-const fixes
  - benchmark_semver.c, test_semver.c, test_uuid.c: split multi-declarations
  - test_semver.c: strcpy -> snprintf for MSVC deprecation
  - test_uuid.c: rephrased comment to not look like commented code

+++ src/thread_pool.c
@@ -278,7 +278,7 @@ size_t zenit_thread_pool_pending(const zenit_thread_pool_t *pool) {
     if (pool == NULL) {
         return 0;
     }
-    return (size_t)(pool->queue_head - pool->queue_tail);
+    return pool->queue_head - pool->queue_tail;
 }
 
 void zenit_thread_pool_wait(zenit_thread_pool_t *pool) {
